Use an org converter directly, saves one step
This bring in an org converter plugin from
https://github.com/bdewey/org-ruby which takes an org file and converts
it to html.

With the proper front-matter, which is the only inconsistency because
that will disturb the orgmode format itself, this eliminates the publish
step from org-mode entirely while keeping the same input method.

This means I do not have to deal with the publishing setup in orgmode
anymore which was giving me problems anyway.