First dirty commit to have a complete orgmode base format
The main thing is that the yaml header can be replaced by key/value
pairs in orgmode syntax so that posts, pages and documents can be valid
orgmode documents in their own right.
This allows to process the org files in other ways in parallel to jekyll
processing them (like having a PDF rendering of the postings)

What works:
- all posts render
- all pages render
- the header is correctly parsed for those two types

What does not work:
- documents are treated differently and do not work completely yet

This needs more work before publishing.